refactor: move otp & users services to libs/resource
