+++ apps/auth/src/auth.module.ts
@@ -2,12 +2,12 @@ import { Module } from '@nestjs/common';
 import { RabbitMQModule, RabbitMQService } from '@app/common';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { UsersModule } from './users/users.module';
+import { UsersModule } from '@app/resource/users';
 import { AccessTokenStrategy, GoogleStrategy, FacebookStrategy } from './strategies';
 import { JwtModule } from '@nestjs/jwt';
 import { JwtGuard } from './guards';
 import { MAIL_SERVICE } from '~/constants';
-import { OtpModule } from './otp';
+import { OtpModule } from '@app/resource/otp';
 
 @Module({
     imports: [
