+++ libs/resource/src/index.ts
@@ -1,0 +1,2 @@
+export * from './users';
+export * from './otp';
