+++ libs/resource/src/otp/index.ts
