+++ libs/resource/src/users/index.ts
@@ -1,0 +1,3 @@
+export * from './users.module';
+export * from './users.service';
+export * from './users.repository';
