+++ libs/resource/src/users/users.module.ts
