+++ package.json
@@ -38,6 +38,7 @@
     "@nestjs/config": "^3.0.0",
     "@nestjs/core": "^10.0.3",
     "@nestjs/jwt": "^10.1.0",
+    "@nestjs/mapped-types": "*",
     "@nestjs/microservices": "^10.0.3",
     "@nestjs/mongoose": "^10.0.0",
     "@nestjs/passport": "^10.0.0",
@@ -125,7 +126,8 @@
       "^~/apps/mail(|/.*)$": "<rootDir>/apps/mail/src/$1",
       "^~/apps/products(|/.*)$": "<rootDir>/apps/products/src/$1",
       "^~/apps/sample(|/.*)$": "<rootDir>/apps/sample/src/$1",
-      "^~/apps/managaments(|/.*)$": "<rootDir>/apps/managaments/src/$1"
+      "^~/apps/managaments(|/.*)$": "<rootDir>/apps/managaments/src/$1",
+      "^@app/resource(|/.*)$": "<rootDir>/libs/resource/src/$1"
     }
   }
 }
