+++ tsconfig.json
@@ -47,6 +47,12 @@
       ],
       "~/apps/managaments/*": [
         "apps/managaments/src/*"
+      ],
+      "@app/resource": [
+        "libs/resource/src"
+      ],
+      "@app/resource/*": [
+        "libs/resource/src/*"
       ]
     }
   }
